ufs2: Fix warning
Reviewed-on: https://review.haiku-os.org/c/haiku/+/6854
Tested-by: Commit checker robot <no-reply+buildbot@haiku-os.org>

+++ src/add-ons/kernel/file_systems/ufs2/DirectoryIterator.cpp
@@ -76,7 +76,7 @@ DirectoryIterator::GetNext(char* name, size_t* _nameLength, ino_t* _id)
 		fOffset += direct.reclen;
 
 		if (direct.next_ino > 0) {
-			if ((direct.namlen + 1) > *_nameLength)
+			if ((size_t) (direct.namlen + 1) > *_nameLength)
 				return B_BUFFER_OVERFLOW;
 			strlcpy(name, direct.name, direct.namlen + 1);
 			*_id = direct.next_ino;
